Add files via upload
Some of the quantum gates. They take a qubit or 2 in matrix form, perform and operation on them, and spit out another qubit or 2. For now, if you want to perform a tensor product on them and then apply them to a statevector, you'll have to do that manually. Currently reworking it so that working with multiple qubits is less of a hassle. Missing SWAP gates and Toffoli gates.